+++ package.json
@@ -111,12 +111,12 @@
         "@babel/plugin-proposal-optional-chaining": "^7.12.7",
         "@pmmmwh/react-refresh-webpack-plugin": "0.5.0-rc.2",
         "@storybook/addon-centered": "^5.3.21",
-        "@storybook/addon-essentials": "^6.3.1",
-        "@storybook/addon-links": "^6.3.1",
-        "@storybook/addons": "^6.3.1",
-        "@storybook/builder-webpack5": "^6.3.1",
-        "@storybook/manager-webpack5": "^6.3.1",
-        "@storybook/react": "^6.3.1",
+        "@storybook/addon-essentials": "^6.5.12",
+        "@storybook/addon-links": "^6.5.12",
+        "@storybook/addons": "^6.5.12",
+        "@storybook/builder-webpack5": "^6.5.12",
+        "@storybook/manager-webpack5": "^6.5.12",
+        "@storybook/react": "^6.5.12",
         "@trivago/prettier-plugin-sort-imports": "^3.2.0",
         "@types/codemirror": "^5.60.5",
         "@types/d3": "^5.7.2",
